requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.2,<6.0
django-cors-headers>=4.0,<5.0
drf-spectacular>=0.26,<1.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
celery>=5.3,<6.0
redis>=4.5,<5.0
django-redis>=5.2,<6.0
python-json-logger>=2.0,<3.0
pytest>=7.3,<8.0
pytest-django>=4.5,<5.0
coverage>=7.2,<8.0
factory-boy>=3.2,<4.0
Faker>=18.0,<19.0
black>=23.3,<24.0
flake8>=6.0,<7.0
isort>=5.12,<6.0
pre-commit>=3.3,<4.0
gunicorn>=20.1,<21.0